Switch c->vfs.inodes to fast_list
The mutex-protected list of bcachefs inodes was a contention point on
every inode load/evict and a serialised walker for the snapshot create
unmap path. Switch to fast_list: lockless add/remove off a percpu slot
buffer, and lockless iteration backed by a genradix.

- pre-allocate the slot index in __bch2_new_inode (the gfp-aware path)
  so the hash-insert success branch is failure-free; release in the
  error paths that bypass evict.
- bch2_evict_inode releases the slot via fast_list_remove (which is a
  no-op when ei_inodes_idx == 0, covering the discard_new_inode race).
- bch2_evict_subvolume_inodes switches to rcu_read_lock +
  fast_list_for_each; inodes are RCU-freed via .free_inode, so the
  pointer is stable until igrab pins or we drop rcu_read_lock.
- bch2_unmap_mapped_inodes in vfs/ioctl.c (snapshot create unmap path)
  uses the same fast_list walker instead of the prior sb-wide
  s_inode_list_lock walker.

inode_state_read* shims move from fs.c to fs.h so ioctl.c can use them.
